requirements.txt
discord.py
imagehash